refactoring ecephys test to generalize with conftest
