+++ allensdk/brain_observatory/__init__.py
@@ -33,3 +33,20 @@
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 #
+from typing import NamedTuple
+
+import numpy as np
+
+
+class RunningSpeed(NamedTuple):
+    ''' Describes the rate at which an experimental subject ran during a session.
+
+    values : np.ndarray
+        running speed (cm/s) at each sample point
+    timestamps : np.ndarray
+        The time at which each sample was collected (s).
+
+    '''
+
+    timestamps: np.ndarray
+    values: np.ndarray
